Common implementation for Lower; remove BinaryOperatorExpression subtypes
